Extract more code block components
